Initialize user account system and update libraries
Refactored the code to integrate an effective user account system. Users can now select different accounts, and undergo the complete authentication process. The update also includes migrating the shared library dependencies to the latest versions. Along with these, couple of redundant files related to the old way account handling logic has been removed and some translation strings are updated to accommodate new feature. Made possible due to changes in schema of MagisterAccount and the update in the logic of fetching accounts. Furthermore, modules 'UtilsAndroid' and 'V21' are updated to handle ByteArray which forms the base of account image processing logic.