Add receiver information and unread messages count
Added a new field 'receiverInfo' to string files for receiver information. Imported necessary packages and declared 'unreadMessages' variable in MenuItems.kt to display the count of unread messages. Updated the 'refreshMessages' function in 'Messages.kt' to refresh messages and to send a notification when a new unread message is received. In 'MagisterAccount.kt', declared methods for refreshing folders and messages. In 'DefaultFolderComponent.kt', calls to refresh folders and messages are made during a folder refresh.

These changes were made to improve the user interface by providing receivers information and to keep users informed about the count of unread messages.